Enable preview button while the editor is disabled
Hi, 

I want to disable the editor as default, so that the user have to enable the editor before he change the content. But I found no way to enable the editor while it's disabled. I think the preview button just changes the content format, therefore, it should be always enabled. on the other hand, if hidePreview function doesn't check disabled and readonly, the other buttons can still work even if the textarea has been disabled which I think is a little bit weird.

Please confirm.