Change preview content height setting
Sets `min-height` rather than `height` and sets `height` to `auto`. This improves the preview look. Container will then never be smaller than textarea but will still adapt to the preview content height. Currently, as the height is hardwired, if the preview content is higher than the textarea, a scrollbar appears.